# docs/swagger.yaml
definitions:
...
  errs.HTTPError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  file.Meta:
    properties:
      category:
        type: string
      filename:
        type: string
      mime:
        type: string
      size:
        type: integer
    required:
    - category
    - mime
    - size
    type: object
  file.ResponseFile:
    properties:
      create_time:
        description: CreateTime holds the value of the "create_time" field.
        type: string
      edges:
        type: object
      filename:
        description: Filename holds the value of the "filename" field.
        type: string
      id:
        description: ID of the ent.
        type: string
      mime:
        description: Mime holds the value of the "mime" field.
        type: string
      note:
        description: Note holds the value of the "note" field.
        type: string
      path:
        description: Path holds the value of the "path" field.
        type: string
      sid:
        description: Sid holds the value of the "sid" field.
        type: string
      size:
        description: Size holds the value of the "size" field.
        type: integer
      status:
        description: Status holds the value of the "status" field.
        type: string
      update_time:
        description: UpdateTime holds the value of the "update_time" field.
        type: string
    type: object
  profile.Profile:
    properties:
      call:
        type: string
      category:
        type: string
      title:
        type: string
    required:
    - call
    - category
    - title
    type: object
  profile.ResponseProfile:
    properties:
      avatar:
        description: Avatar holds the value of the "avatar" field.
        type: string
      call:
        description: Call holds the value of the "call" field.
        type: string
      category:
        description: Category holds the value of the "category" field.
        type: string
      create_time:
        description: CreateTime holds the value of the "create_time" field.
        type: string
      edges:
        type: object
      id:
        description: ID of the ent.
        type: string
      sid:
        description: Sid holds the value of the "sid" field.
        type: string
      title:
        description: Title holds the value of the "title" field.
        type: string
      update_time:
        description: UpdateTime holds the value of the "update_time" field.
        type: string
    type: object
  series.ResponseSeries:
    properties:
      call:
        description: Call holds the value of the "call" field.
        type: string
      content:
        description: Content holds the value of the "content" field.
        type: string
      create_time:
        description: CreateTime holds the value of the "create_time" field.
        type: string
      edges:
        type: object
      id:
        description: ID of the ent.
        type: string
      sid:
        description: Sid holds the value of the "sid" field.
        type: string
      title:
        description: Title holds the value of the "title" field.
        type: string
      update_time:
        description: UpdateTime holds the value of the "update_time" field.
        type: string
    type: object
  series.TitleContent:
    properties:
      content:
        type: string
      title:
        type: string
    required:
    - title
    type: object
info:
  contact: {}
paths:
...
  /file/large/{pid}:
    post:
      consumes:
      - application/json
      parameters:
      - description: profile id
        in: path
        name: pid
        required: true
        type: string
      - description: file meta
        in: body
        name: meta
        required: true
        schema:
          $ref: '#/definitions/file.Meta'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/file.ResponseFile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.HTTPError'
      summary: create multipart upload
      tags:
      - file
  /file/large/{pid}/{id}:
    post:
      parameters:
      - description: profile id
        in: path
        name: pid
        required: true
        type: string
      - description: file id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/file.ResponseFile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.HTTPError'
      summary: finish multipart upload
      tags:
      - file
  /file/large/{pid}/{id}/{part_id}:
    put:
      consumes:
      - application/octet-stream
      parameters:
      - description: profile id
        in: path
        name: pid
        required: true
        type: string
      - description: file id
        in: path
        name: id
        required: true
        type: string
      - description: part id
        in: path
        name: part_id
        required: true
        type: integer
      - description: file content
        in: body
        name: data
        required: true
        schema:
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.HTTPError'
      summary: create multipart upload
      tags:
      - file
  /file/single/{pid}:
    post:
      consumes:
      - application/json
      parameters:
      - description: profile id
        in: path
        name: pid
        required: true
        type: string
      - description: file meta
        in: body
        name: meta
        required: true
        schema:
          $ref: '#/definitions/file.Meta'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/file.ResponseFile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.HTTPError'
      summary: create single file upload
      tags:
      - file
  /file/single/{pid}/{id}:
    put:
      consumes:
      - application/octet-stream
      parameters:
      - description: profile id
        in: path
        name: pid
        required: true
        type: string
      - description: file id
        in: path
        name: id
        required: true
        type: string
      - description: file content
        in: body
        name: data
        required: true
        schema:
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/file.ResponseFile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.HTTPError'
      summary: single file content upload
      tags:
      - file
  /profile:
    post:
      consumes:
      - application/json
      parameters:
      - description: profile info
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/profile.Profile'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/profile.ResponseProfile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.HTTPError'
      summary: create profile
      tags:
      - profile
  /profile/{pid}:
    get:
      parameters:
      - description: profile id
        in: path
        name: pid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/profile.ResponseProfile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.HTTPError'
      summary: get profile
      tags:
      - profile
    put:
      consumes:
      - application/json
      parameters:
      - description: profile id
        in: path
        name: pid
        required: true
        type: string
      - description: profile info
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/profile.Profile'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/profile.ResponseProfile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.HTTPError'
      summary: update profile
      tags:
      - profile
  /series/{id}:
    get:
      parameters:
      - description: series id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/series.ResponseSeries'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.HTTPError'
      summary: get a series
      tags:
      - series
  /series/{pid}:
    post:
      consumes:
      - application/json
      parameters:
      - description: profile id
        in: path
        name: pid
        required: true
        type: string
      - description: series info
        in: body
        name: series
        required: true
        schema:
          $ref: '#/definitions/series.TitleContent'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/series.ResponseSeries'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.HTTPError'
      summary: create a series
      tags:
      - series
  /series/{pid}/{id}:
    delete:
      parameters:
      - description: profile id
        in: path
        name: pid
        required: true
        type: string
      - description: series id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.HTTPError'
      summary: del a series
      tags:
      - series
    put:
      consumes:
      - application/json
      parameters:
      - description: series info
        in: body
        name: series
        required: true
        schema:
          $ref: '#/definitions/series.TitleContent'
      - description: profile id
        in: path
        name: pid
        required: true
        type: string
      - description: series id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/series.ResponseSeries'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.HTTPError'
      summary: update a series
      tags:
      - series
swagger: "2.0"